Move errorReplacer into redact util
